pyproject.toml
[project]
name = "minio-scripts"
version = "0.1.0"
description = ""
authors = [
    {name = "nyktyz",email = "krovtrusa@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "minio (>=7.2.19,<8.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
